Create manager class to isolate logic for ciphers (#1432)
