+++ app/src/test/java/com/x8bit/bitwarden/data/vault/repository/VaultRepositoryTest.kt
@@ -3,7 +3,6 @@ package com.x8bit.bitwarden.data.vault.repository
 import android.net.Uri
 import app.cash.turbine.test
 import app.cash.turbine.turbineScope
-import com.bitwarden.core.Attachment
 import com.bitwarden.core.Cipher
 import com.bitwarden.core.CipherView
 import com.bitwarden.core.CollectionView
@@ -37,21 +36,14 @@ import com.x8bit.bitwarden.data.platform.repository.util.bufferedMutableSharedFl
 import com.x8bit.bitwarden.data.platform.util.asFailure
 import com.x8bit.bitwarden.data.platform.util.asSuccess
 import com.x8bit.bitwarden.data.vault.datasource.disk.VaultDiskSource
-import com.x8bit.bitwarden.data.vault.datasource.network.model.AttachmentJsonRequest
-import com.x8bit.bitwarden.data.vault.datasource.network.model.CreateCipherInOrganizationJsonRequest
 import com.x8bit.bitwarden.data.vault.datasource.network.model.CreateFileSendResponse
 import com.x8bit.bitwarden.data.vault.datasource.network.model.CreateSendJsonResponse
 import com.x8bit.bitwarden.data.vault.datasource.network.model.FolderJsonRequest
 import com.x8bit.bitwarden.data.vault.datasource.network.model.SendTypeJson
-import com.x8bit.bitwarden.data.vault.datasource.network.model.ShareCipherJsonRequest
 import com.x8bit.bitwarden.data.vault.datasource.network.model.SyncResponseJson
-import com.x8bit.bitwarden.data.vault.datasource.network.model.UpdateCipherCollectionsJsonRequest
-import com.x8bit.bitwarden.data.vault.datasource.network.model.UpdateCipherResponseJson
 import com.x8bit.bitwarden.data.vault.datasource.network.model.UpdateFolderResponseJson
 import com.x8bit.bitwarden.data.vault.datasource.network.model.UpdateSendResponseJson
-import com.x8bit.bitwarden.data.vault.datasource.network.model.createMockAttachmentJsonResponse
 import com.x8bit.bitwarden.data.vault.datasource.network.model.createMockCipher
-import com.x8bit.bitwarden.data.vault.datasource.network.model.createMockCipherJsonRequest
 import com.x8bit.bitwarden.data.vault.datasource.network.model.createMockCollection
 import com.x8bit.bitwarden.data.vault.datasource.network.model.createMockDomains
 import com.x8bit.bitwarden.data.vault.datasource.network.model.createMockFileSendResponseJson
@@ -69,38 +61,28 @@ import com.x8bit.bitwarden.data.vault.datasource.network.service.SendsService
 import com.x8bit.bitwarden.data.vault.datasource.network.service.SyncService
 import com.x8bit.bitwarden.data.vault.datasource.sdk.VaultSdkSource
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.InitializeCryptoResult
-import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockAttachmentView
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockCipherView
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockCollectionView
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockFolderView
-import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockSdkAttachment
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockSdkCipher
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockSdkCollection
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockSdkFolder
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockSdkSend
 import com.x8bit.bitwarden.data.vault.datasource.sdk.model.createMockSendView
+import com.x8bit.bitwarden.data.vault.manager.CipherManager
 import com.x8bit.bitwarden.data.vault.manager.FileManager
 import com.x8bit.bitwarden.data.vault.manager.TotpCodeManager
 import com.x8bit.bitwarden.data.vault.manager.VaultLockManager
-import com.x8bit.bitwarden.data.vault.manager.model.DownloadResult
 import com.x8bit.bitwarden.data.vault.manager.model.VerificationCodeItem
-import com.x8bit.bitwarden.data.vault.repository.model.CreateAttachmentResult
-import com.x8bit.bitwarden.data.vault.repository.model.CreateCipherResult
 import com.x8bit.bitwarden.data.vault.repository.model.CreateFolderResult
 import com.x8bit.bitwarden.data.vault.repository.model.CreateSendResult
-import com.x8bit.bitwarden.data.vault.repository.model.DeleteAttachmentResult
-import com.x8bit.bitwarden.data.vault.repository.model.DeleteCipherResult
 import com.x8bit.bitwarden.data.vault.repository.model.DeleteFolderResult
 import com.x8bit.bitwarden.data.vault.repository.model.DeleteSendResult
 import com.x8bit.bitwarden.data.vault.repository.model.DomainsData
-import com.x8bit.bitwarden.data.vault.repository.model.DownloadAttachmentResult
 import com.x8bit.bitwarden.data.vault.repository.model.ExportVaultDataResult
 import com.x8bit.bitwarden.data.vault.repository.model.GenerateTotpResult
 import com.x8bit.bitwarden.data.vault.repository.model.RemovePasswordSendResult
-import com.x8bit.bitwarden.data.vault.repository.model.RestoreCipherResult
 import com.x8bit.bitwarden.data.vault.repository.model.SendData
-import com.x8bit.bitwarden.data.vault.repository.model.ShareCipherResult
-import com.x8bit.bitwarden.data.vault.repository.model.UpdateCipherResult
 import com.x8bit.bitwarden.data.vault.repository.model.UpdateFolderResult
 import com.x8bit.bitwarden.data.vault.repository.model.UpdateSendResult
 import com.x8bit.bitwarden.data.vault.repository.model.VaultData
@@ -185,6 +167,7 @@ class VaultRepositoryTest {
         emptyList(),
     )
     private val mutableUnlockedUserIdsStateFlow = MutableStateFlow<Set<String>>(emptySet())
+    private val cipherManager: CipherManager = mockk()
     private val vaultLockManager: VaultLockManager = mockk {
         every { vaultUnlockDataStateFlow } returns mutableVaultStateFlow
         every {
@@ -234,6 +217,7 @@ class VaultRepositoryTest {
         dispatcherManager = dispatcherManager,
         totpCodeManager = totpCodeManager,
         pushManager = pushManager,
+        cipherManager = cipherManager,
         fileManager = fileManager,
         clock = clock,
         userLogoutManager = userLogoutManager,
@@ -1909,637 +1893,6 @@ class VaultRepositoryTest {
         }
 
     @Test
-    fun `createCipher with no active user should return CreateCipherResult failure`() =
-        runTest {
-            fakeAuthDiskSource.userState = null
-
-            val result = vaultRepository.createCipher(cipherView = mockk())
-
-            assertEquals(
-                CreateCipherResult.Error,
-                result,
-            )
-        }
-
-    @Test
-    fun `createCipher with encryptCipher failure should return CreateCipherResult failure`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val mockCipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = mockCipherView,
-                )
-            } returns IllegalStateException().asFailure()
-
-            val result = vaultRepository.createCipher(cipherView = mockCipherView)
-
-            assertEquals(
-                CreateCipherResult.Error,
-                result,
-            )
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `createCipher with ciphersService createCipher failure should return CreateCipherResult failure`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val mockCipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = mockCipherView,
-                )
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            coEvery {
-                ciphersService.createCipher(
-                    body = createMockCipherJsonRequest(number = 1, hasNullUri = true),
-                )
-            } returns IllegalStateException().asFailure()
-
-            val result = vaultRepository.createCipher(cipherView = mockCipherView)
-
-            assertEquals(
-                CreateCipherResult.Error,
-                result,
-            )
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `createCipher with ciphersService createCipher success should return CreateCipherResult success`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val mockCipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = mockCipherView,
-                )
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            val mockCipher = createMockCipher(number = 1)
-            coEvery {
-                ciphersService.createCipher(
-                    body = createMockCipherJsonRequest(number = 1, hasNullUri = true),
-                )
-            } returns mockCipher.asSuccess()
-            coEvery { vaultDiskSource.saveCipher(userId, mockCipher) } just runs
-
-            val result = vaultRepository.createCipher(cipherView = mockCipherView)
-
-            assertEquals(
-                CreateCipherResult.Success,
-                result,
-            )
-        }
-
-    @Test
-    fun `createCipherInOrganization with no active user should return CreateCipherResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = null
-
-            val result = vaultRepository.createCipherInOrganization(
-                cipherView = mockk(),
-                collectionIds = mockk(),
-            )
-
-            assertEquals(
-                CreateCipherResult.Error,
-                result,
-            )
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `createCipherInOrganization with encryptCipher failure should return CreateCipherResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val mockCipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = mockCipherView,
-                )
-            } returns IllegalStateException().asFailure()
-
-            val result = vaultRepository.createCipherInOrganization(
-                cipherView = mockCipherView,
-                collectionIds = mockk(),
-            )
-
-            assertEquals(
-                CreateCipherResult.Error,
-                result,
-            )
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `createCipherInOrganization with ciphersService createCipher failure should return CreateCipherResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val mockCipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = mockCipherView,
-                )
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            coEvery {
-                ciphersService.createCipherInOrganization(
-                    body = CreateCipherInOrganizationJsonRequest(
-                        cipher = createMockCipherJsonRequest(number = 1, hasNullUri = true),
-                        collectionIds = listOf("mockId-1"),
-                    ),
-                )
-            } returns IllegalStateException().asFailure()
-
-            val result = vaultRepository.createCipherInOrganization(
-                cipherView = mockCipherView,
-                collectionIds = listOf("mockId-1"),
-            )
-
-            assertEquals(
-                CreateCipherResult.Error,
-                result,
-            )
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `createCipherInOrganization with ciphersService createCipher success should return CreateCipherResult success`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val mockCipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = mockCipherView,
-                )
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            val mockCipher = createMockCipher(number = 1)
-            coEvery {
-                ciphersService.createCipherInOrganization(
-                    body = CreateCipherInOrganizationJsonRequest(
-                        cipher = createMockCipherJsonRequest(number = 1, hasNullUri = true),
-                        collectionIds = listOf("mockId-1"),
-                    ),
-                )
-            } returns mockCipher.asSuccess()
-            coEvery {
-                vaultDiskSource.saveCipher(
-                    userId,
-                    mockCipher.copy(collectionIds = listOf("mockId-1")),
-                )
-            } just runs
-
-            val result = vaultRepository.createCipherInOrganization(
-                cipherView = mockCipherView,
-                collectionIds = listOf("mockId-1"),
-            )
-
-            assertEquals(
-                CreateCipherResult.Success,
-                result,
-            )
-        }
-
-    @Test
-    fun `updateCipher with no active user should return UpdateCipherResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = null
-
-            val result = vaultRepository.updateCipher(
-                cipherId = "cipherId",
-                cipherView = mockk(),
-            )
-
-            assertEquals(
-                UpdateCipherResult.Error(null),
-                result,
-            )
-        }
-
-    @Test
-    fun `updateCipher with encryptCipher failure should return UpdateCipherResult failure`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "cipherId1234"
-            val mockCipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = mockCipherView,
-                )
-            } returns IllegalStateException().asFailure()
-
-            val result = vaultRepository.updateCipher(
-                cipherId = cipherId,
-                cipherView = mockCipherView,
-            )
-
-            assertEquals(UpdateCipherResult.Error(errorMessage = null), result)
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `updateCipher with ciphersService updateCipher failure should return UpdateCipherResult Error with a null message`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "cipherId1234"
-            val mockCipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = mockCipherView,
-                )
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            coEvery {
-                ciphersService.updateCipher(
-                    cipherId = cipherId,
-                    body = createMockCipherJsonRequest(number = 1, hasNullUri = true),
-                )
-            } returns IllegalStateException().asFailure()
-
-            val result = vaultRepository.updateCipher(
-                cipherId = cipherId,
-                cipherView = mockCipherView,
-            )
-
-            assertEquals(UpdateCipherResult.Error(errorMessage = null), result)
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `updateCipher with ciphersService updateCipher Invalid response should return UpdateCipherResult Error with a non-null message`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "cipherId1234"
-            val mockCipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = mockCipherView,
-                )
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            coEvery {
-                ciphersService.updateCipher(
-                    cipherId = cipherId,
-                    body = createMockCipherJsonRequest(number = 1, hasNullUri = true),
-                )
-            } returns UpdateCipherResponseJson
-                .Invalid(
-                    message = "You do not have permission to edit this.",
-                    validationErrors = null,
-                )
-                .asSuccess()
-
-            val result = vaultRepository.updateCipher(
-                cipherId = cipherId,
-                cipherView = mockCipherView,
-            )
-
-            assertEquals(
-                UpdateCipherResult.Error(
-                    errorMessage = "You do not have permission to edit this.",
-                ),
-                result,
-            )
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `updateCipher with ciphersService updateCipher Success response should return UpdateCipherResult success`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "cipherId1234"
-            val mockCipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = mockCipherView,
-                )
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            val mockCipher = createMockCipher(number = 1)
-            coEvery {
-                ciphersService.updateCipher(
-                    cipherId = cipherId,
-                    body = createMockCipherJsonRequest(number = 1, hasNullUri = true),
-                )
-            } returns UpdateCipherResponseJson
-                .Success(cipher = mockCipher)
-                .asSuccess()
-            coEvery {
-                vaultDiskSource.saveCipher(
-                    userId = userId,
-                    cipher = mockCipher.copy(collectionIds = mockCipherView.collectionIds),
-                )
-            } just runs
-
-            val result = vaultRepository.updateCipher(
-                cipherId = cipherId,
-                cipherView = mockCipherView,
-            )
-
-            assertEquals(UpdateCipherResult.Success, result)
-        }
-
-    @Test
-    fun `hardDeleteCipher with no active user should return DeleteCipherResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = null
-
-            val result = vaultRepository.hardDeleteCipher(
-                cipherId = "cipherId",
-            )
-
-            assertEquals(
-                DeleteCipherResult.Error,
-                result,
-            )
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `hardDeleteCipher with ciphersService hardDeleteCipher failure should return DeleteCipherResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val cipherId = "mockId-1"
-            coEvery {
-                ciphersService.hardDeleteCipher(cipherId = cipherId)
-            } returns Throwable("Fail").asFailure()
-
-            val result = vaultRepository.hardDeleteCipher(cipherId)
-
-            assertEquals(DeleteCipherResult.Error, result)
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `hardDeleteCipher with ciphersService hardDeleteCipher success should return DeleteCipherResult success`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "mockId-1"
-            coEvery { ciphersService.hardDeleteCipher(cipherId = cipherId) } returns Unit.asSuccess()
-            coEvery { vaultDiskSource.deleteCipher(userId, cipherId) } just runs
-
-            val result = vaultRepository.hardDeleteCipher(cipherId)
-
-            assertEquals(DeleteCipherResult.Success, result)
-        }
-
-    @Test
-    fun `softDeleteCipher with no active user should return DeleteCipherResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = null
-
-            val result = vaultRepository.softDeleteCipher(
-                cipherId = "cipherId",
-                cipherView = mockk(),
-            )
-
-            assertEquals(
-                DeleteCipherResult.Error,
-                result,
-            )
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `softDeleteCipher with ciphersService softDeleteCipher failure should return DeleteCipherResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val cipherId = "mockId-1"
-            coEvery {
-                ciphersService.softDeleteCipher(cipherId = cipherId)
-            } returns Throwable("Fail").asFailure()
-
-            val result = vaultRepository.softDeleteCipher(
-                cipherId = cipherId,
-                cipherView = createMockCipherView(number = 1),
-            )
-
-            assertEquals(DeleteCipherResult.Error, result)
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `softDeleteCipher with ciphersService softDeleteCipher success should return DeleteCipherResult success`() =
-        runTest {
-            mockkStatic(Cipher::toEncryptedNetworkCipherResponse)
-            every {
-                createMockSdkCipher(number = 1, clock = clock).toEncryptedNetworkCipherResponse()
-            } returns createMockCipher(number = 1)
-            val fixedInstant = Instant.parse("2021-01-01T00:00:00Z")
-            val userId = "mockId-1"
-            val cipherId = "mockId-1"
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = createMockCipherView(number = 1)
-                        .copy(
-                            deletedDate = fixedInstant,
-                        ),
-                )
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            coEvery { ciphersService.softDeleteCipher(cipherId = cipherId) } returns Unit.asSuccess()
-            coEvery {
-                vaultDiskSource.saveCipher(
-                    userId = userId,
-                    cipher = createMockCipher(number = 1),
-                )
-            } just runs
-            val cipherView = createMockCipherView(number = 1)
-            mockkStatic(Instant::class)
-            every { Instant.now() } returns fixedInstant
-
-            val result = vaultRepository.softDeleteCipher(
-                cipherId = cipherId,
-                cipherView = cipherView,
-            )
-
-            assertEquals(DeleteCipherResult.Success, result)
-            unmockkStatic(Instant::class)
-            unmockkStatic(Cipher::toEncryptedNetworkCipherResponse)
-        }
-
-    @Test
-    fun `deleteCipherAttachment with no active user should return DeleteAttachmentResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = null
-
-            val result = vaultRepository.deleteCipherAttachment(
-                cipherId = "cipherId",
-                attachmentId = "attachmentId",
-                cipherView = mockk(),
-            )
-
-            assertEquals(
-                DeleteAttachmentResult.Error,
-                result,
-            )
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `deleteCipherAttachment with ciphersService deleteCipherAttachment failure should return DeleteAttachmentResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val cipherId = "mockId-1"
-            val attachmentId = "mockId-1"
-            coEvery {
-                ciphersService.deleteCipherAttachment(
-                    cipherId = cipherId,
-                    attachmentId = attachmentId,
-                )
-            } returns Throwable("Fail").asFailure()
-
-            val result = vaultRepository.deleteCipherAttachment(
-                cipherId = cipherId,
-                attachmentId = attachmentId,
-                cipherView = createMockCipherView(number = 1),
-            )
-
-            assertEquals(DeleteAttachmentResult.Error, result)
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `deleteCipherAttachment with ciphersService deleteCipherAttachment success should return DeleteAttachmentResult success`() =
-        runTest {
-            mockkStatic(Cipher::toEncryptedNetworkCipherResponse)
-            every {
-                createMockSdkCipher(number = 1, clock = clock).toEncryptedNetworkCipherResponse()
-            } returns createMockCipher(number = 1)
-            val fixedInstant = Instant.parse("2021-01-01T00:00:00Z")
-            val userId = "mockId-1"
-            val cipherId = "mockId-1"
-            val attachmentId = "mockId-1"
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = createMockCipherView(number = 1).copy(
-                        attachments = emptyList(),
-                    ),
-                )
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            coEvery {
-                ciphersService.deleteCipherAttachment(
-                    cipherId = cipherId,
-                    attachmentId = attachmentId,
-                )
-            } returns Unit.asSuccess()
-            coEvery {
-                vaultDiskSource.saveCipher(
-                    userId = userId,
-                    cipher = createMockCipher(number = 1),
-                )
-            } just runs
-            val cipherView = createMockCipherView(number = 1)
-            mockkStatic(Instant::class)
-            every { Instant.now() } returns fixedInstant
-
-            val result = vaultRepository.deleteCipherAttachment(
-                cipherId = cipherId,
-                attachmentId = attachmentId,
-                cipherView = cipherView,
-            )
-
-            assertEquals(DeleteAttachmentResult.Success, result)
-            unmockkStatic(Instant::class)
-            unmockkStatic(Cipher::toEncryptedNetworkCipherResponse)
-        }
-
-    @Test
-    fun `restoreCipher with no active user should return RestoreCipherResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = null
-
-            val result = vaultRepository.restoreCipher(
-                cipherId = "cipherId",
-                cipherView = mockk(),
-            )
-
-            assertEquals(
-                RestoreCipherResult.Error,
-                result,
-            )
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `restoreCipher with ciphersService restoreCipher failure should return RestoreCipherResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val cipherId = "mockId-1"
-            coEvery {
-                ciphersService.restoreCipher(cipherId = cipherId)
-            } returns Throwable("Fail").asFailure()
-
-            val result = vaultRepository.restoreCipher(
-                cipherId = cipherId,
-                cipherView = createMockCipherView(number = 1),
-            )
-
-            assertEquals(RestoreCipherResult.Error, result)
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `restoreCipher with ciphersService restoreCipher success should return RestoreCipherResult success`() =
-        runTest {
-            mockkStatic(Cipher::toEncryptedNetworkCipherResponse)
-            every {
-                createMockSdkCipher(number = 1, clock = clock).toEncryptedNetworkCipherResponse()
-            } returns createMockCipher(number = 1)
-            val fixedInstant = Instant.parse("2021-01-01T00:00:00Z")
-            val userId = "mockId-1"
-            val cipherId = "mockId-1"
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = createMockCipherView(number = 1)
-                        .copy(
-                            deletedDate = null,
-                        ),
-                )
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            coEvery { ciphersService.restoreCipher(cipherId = cipherId) } returns Unit.asSuccess()
-            coEvery {
-                vaultDiskSource.saveCipher(
-                    userId = userId,
-                    cipher = createMockCipher(number = 1),
-                )
-            } just runs
-            val cipherView = createMockCipherView(number = 1)
-            mockkStatic(Instant::class)
-            every { Instant.now() } returns fixedInstant
-
-            val result = vaultRepository.restoreCipher(
-                cipherId = cipherId,
-                cipherView = cipherView,
-            )
-
-            assertEquals(RestoreCipherResult.Success, result)
-        }
-
-    @Test
     fun `createSend with no active user should return CreateSendResult Error`() =
         runTest {
             fakeAuthDiskSource.userState = null
@@ -3085,1116 +2438,6 @@ class VaultRepositoryTest {
             val result = vaultRepository.deleteSend(sendId)
 
             assertEquals(DeleteSendResult.Success, result)
-        }
-
-    @Test
-    fun `shareCipher with no active user should return ShareCipherResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = null
-
-            val result = vaultRepository.shareCipher(
-                cipherId = "cipherId",
-                organizationId = "organizationId",
-                cipherView = mockk(),
-                collectionIds = emptyList(),
-            )
-
-            assertEquals(
-                ShareCipherResult.Error,
-                result,
-            )
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `shareCipher with cipherService shareCipher success should return ShareCipherResultSuccess`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val organizationId = "organizationId"
-            val mockCipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.moveToOrganization(
-                    userId = userId,
-                    organizationId = organizationId,
-                    cipherView = createMockCipherView(number = 1),
-                )
-            } returns mockCipherView.asSuccess()
-            coEvery {
-                vaultSdkSource.encryptCipher(userId = userId, cipherView = mockCipherView)
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            coEvery {
-                ciphersService.shareCipher(
-                    cipherId = "mockId-1",
-                    body = ShareCipherJsonRequest(
-                        cipher = createMockCipherJsonRequest(number = 1, hasNullUri = true),
-                        collectionIds = listOf("mockId-1"),
-                    ),
-                )
-            } returns createMockCipher(number = 1).asSuccess()
-            coEvery {
-                vaultDiskSource.saveCipher(
-                    userId,
-                    createMockCipher(number = 1)
-                        .copy(collectionIds = listOf("mockId-1")),
-                )
-            } just runs
-
-            val result = vaultRepository.shareCipher(
-                cipherId = "mockId-1",
-                organizationId = organizationId,
-                cipherView = createMockCipherView(number = 1),
-                collectionIds = listOf("mockId-1"),
-            )
-
-            assertEquals(
-                ShareCipherResult.Success,
-                result,
-            )
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `shareCipher with cipherService shareCipher failure should return ShareCipherResultError`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val organizationId = "organizationId"
-            val mockCipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.moveToOrganization(
-                    userId = userId,
-                    organizationId = organizationId,
-                    cipherView = createMockCipherView(number = 1),
-                )
-            } returns mockCipherView.asSuccess()
-            coEvery {
-                vaultSdkSource.encryptCipher(userId = userId, cipherView = mockCipherView)
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            coEvery {
-                ciphersService.shareCipher(
-                    cipherId = "mockId-1",
-                    body = ShareCipherJsonRequest(
-                        cipher = createMockCipherJsonRequest(number = 1, hasNullUri = true),
-                        collectionIds = listOf("mockId-1"),
-                    ),
-                )
-            } returns Throwable("Fail").asFailure()
-            coEvery { vaultDiskSource.saveCipher(userId, createMockCipher(number = 1)) } just runs
-
-            val result = vaultRepository.shareCipher(
-                cipherId = "mockId-1",
-                organizationId = organizationId,
-                cipherView = createMockCipherView(number = 1),
-                collectionIds = listOf("mockId-1"),
-            )
-
-            assertEquals(
-                ShareCipherResult.Error,
-                result,
-            )
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `shareCipher with cipherService encryptCipher failure should return ShareCipherResultError`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val organizationId = "organizationId"
-            coEvery {
-                vaultSdkSource.moveToOrganization(
-                    userId = userId,
-                    organizationId = organizationId,
-                    cipherView = createMockCipherView(number = 1),
-                )
-            } returns Throwable("Fail").asFailure()
-            coEvery {
-                ciphersService.shareCipher(
-                    cipherId = "mockId-1",
-                    body = ShareCipherJsonRequest(
-                        cipher = createMockCipherJsonRequest(number = 1),
-                        collectionIds = listOf("mockId-1"),
-                    ),
-                )
-            } returns createMockCipher(number = 1).asSuccess()
-            coEvery { vaultDiskSource.saveCipher(userId, createMockCipher(number = 1)) } just runs
-
-            val result = vaultRepository.shareCipher(
-                cipherId = "mockId-1",
-                organizationId = organizationId,
-                cipherView = createMockCipherView(number = 1),
-                collectionIds = listOf("mockId-1"),
-            )
-
-            assertEquals(
-                ShareCipherResult.Error,
-                result,
-            )
-        }
-
-    @Test
-    fun `updateCipherCollections with no active user should return ShareCipherResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = null
-
-            val result = vaultRepository.updateCipherCollections(
-                cipherId = "cipherId",
-                cipherView = mockk(),
-                collectionIds = emptyList(),
-            )
-
-            assertEquals(
-                ShareCipherResult.Error,
-                result,
-            )
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `updateCipherCollections with cipherService updateCipherCollections success should return ShareCipherResultSuccess`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = createMockCipherView(number = 1),
-                )
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            coEvery {
-                ciphersService.updateCipherCollections(
-                    cipherId = "mockId-1",
-                    body = UpdateCipherCollectionsJsonRequest(
-                        collectionIds = listOf("mockId-1"),
-                    ),
-                )
-            } returns Unit.asSuccess()
-            coEvery { vaultDiskSource.saveCipher(userId, any()) } just runs
-
-            val result = vaultRepository.updateCipherCollections(
-                cipherId = "mockId-1",
-                cipherView = createMockCipherView(number = 1)
-                    .copy(collectionIds = listOf("mockId-1")),
-                collectionIds = listOf("mockId-1"),
-            )
-
-            assertEquals(
-                ShareCipherResult.Success,
-                result,
-            )
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `updateCipherCollections with updateCipherCollections shareCipher failure should return ShareCipherResultError`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = createMockCipherView(number = 1),
-                )
-            } returns createMockSdkCipher(number = 1, clock = clock).asSuccess()
-            coEvery {
-                ciphersService.updateCipherCollections(
-                    cipherId = "mockId-1",
-                    body = UpdateCipherCollectionsJsonRequest(
-                        collectionIds = listOf("mockId-1"),
-                    ),
-                )
-            } returns Throwable("Fail").asFailure()
-            coEvery { vaultDiskSource.saveCipher(userId, any()) } just runs
-
-            val result = vaultRepository.updateCipherCollections(
-                cipherId = "mockId-1",
-                cipherView = createMockCipherView(number = 1)
-                    .copy(collectionIds = listOf("mockId-1")),
-                collectionIds = listOf("mockId-1"),
-            )
-
-            assertEquals(
-                ShareCipherResult.Error,
-                result,
-            )
-        }
-
-    @Test
-    @Suppress("MaxLineLength")
-    fun `updateCipherCollections with updateCipherCollections encryptCipher failure should return ShareCipherResultError`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            coEvery {
-                vaultSdkSource.encryptCipher(
-                    userId = userId,
-                    cipherView = createMockCipherView(number = 1),
-                )
-            } returns Throwable("Fail").asFailure()
-            coEvery {
-                ciphersService.updateCipherCollections(
-                    cipherId = "mockId-1",
-                    body = UpdateCipherCollectionsJsonRequest(
-                        collectionIds = listOf("mockId-1"),
-                    ),
-                )
-            } returns Unit.asSuccess()
-            coEvery { vaultDiskSource.saveCipher(userId, any()) } just runs
-
-            val result = vaultRepository.updateCipherCollections(
-                cipherId = "mockId-1",
-                cipherView = createMockCipherView(number = 1)
-                    .copy(collectionIds = listOf("mockId-1")),
-                collectionIds = listOf("mockId-1"),
-            )
-
-            assertEquals(
-                ShareCipherResult.Error,
-                result,
-            )
-        }
-
-    @Test
-    fun `createAttachment with no active user should return CreateAttachmentResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = null
-
-            val result = vaultRepository.createAttachment(
-                cipherId = "cipherId",
-                cipherView = mockk(),
-                fileSizeBytes = "mockFileSize",
-                fileName = "mockFileName",
-                fileUri = mockk(),
-            )
-
-            assertEquals(
-                CreateAttachmentResult.Error,
-                result,
-            )
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `createAttachment with encryptCipher failure should return CreateAttachmentResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "cipherId-1"
-            val mockUri = setupMockUri(url = "www.test.com")
-            val mockCipherView = createMockCipherView(number = 1)
-            val mockFileName = "mockFileName-1"
-            val mockFileSize = "1"
-            coEvery {
-                vaultSdkSource.encryptCipher(userId = userId, cipherView = mockCipherView)
-            } returns Throwable("Fail").asFailure()
-
-            val result = vaultRepository.createAttachment(
-                cipherId = cipherId,
-                cipherView = mockCipherView,
-                fileSizeBytes = mockFileSize,
-                fileName = mockFileName,
-                fileUri = mockUri,
-            )
-
-            assertEquals(CreateAttachmentResult.Error, result)
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `createAttachment with encryptAttachment failure should return CreateAttachmentResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "cipherId-1"
-            val mockUri = setupMockUri(url = "www.test.com")
-            val mockCipherView = createMockCipherView(number = 1)
-            val mockCipher = createMockSdkCipher(number = 1, clock = clock)
-            val mockFile = File.createTempFile("mockFile", "temp")
-            val mockFileName = "mockFileName-1"
-            val mockFileSize = "1"
-            val mockAttachmentView = createMockAttachmentView(number = 1).copy(
-                sizeName = null,
-                id = null,
-                url = null,
-                key = null,
-            )
-            coEvery {
-                vaultSdkSource.encryptCipher(userId = userId, cipherView = mockCipherView)
-            } returns mockCipher.asSuccess()
-            coEvery {
-                fileManager.writeUriToCache(fileUri = mockUri)
-            } returns mockFile.asSuccess()
-            coEvery {
-                vaultSdkSource.encryptAttachment(
-                    userId = userId,
-                    cipher = mockCipher,
-                    attachmentView = mockAttachmentView,
-                    decryptedFilePath = mockFile.absolutePath,
-                    encryptedFilePath = "${mockFile.absolutePath}.enc",
-                )
-            } returns Throwable("Fail").asFailure()
-
-            val result = vaultRepository.createAttachment(
-                cipherId = cipherId,
-                cipherView = mockCipherView,
-                fileSizeBytes = mockFileSize,
-                fileName = mockFileName,
-                fileUri = mockUri,
-            )
-
-            assertEquals(CreateAttachmentResult.Error, result)
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `createAttachment with uriToByteArray failure should return CreateAttachmentResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "cipherId-1"
-            val mockUri = setupMockUri(url = "www.test.com")
-            val mockCipherView = createMockCipherView(number = 1)
-            val mockCipher = createMockSdkCipher(number = 1, clock = clock)
-            val mockFileName = "mockFileName-1"
-            val mockFileSize = "1"
-            coEvery {
-                vaultSdkSource.encryptCipher(userId = userId, cipherView = mockCipherView)
-            } returns mockCipher.asSuccess()
-            coEvery {
-                fileManager.writeUriToCache(fileUri = mockUri)
-            } returns Throwable("Fail").asFailure()
-
-            val result = vaultRepository.createAttachment(
-                cipherId = cipherId,
-                cipherView = mockCipherView,
-                fileSizeBytes = mockFileSize,
-                fileName = mockFileName,
-                fileUri = mockUri,
-            )
-
-            assertEquals(CreateAttachmentResult.Error, result)
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `createAttachment with createAttachment failure should return CreateAttachmentResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "cipherId-1"
-            val mockUri = setupMockUri(url = "www.test.com")
-            val mockCipherView = createMockCipherView(number = 1)
-            val mockCipher = createMockSdkCipher(number = 1, clock = clock)
-            val mockFileName = "mockFileName-1"
-            val mockFileSize = "1"
-            val mockAttachmentView = createMockAttachmentView(number = 1).copy(
-                sizeName = null,
-                id = null,
-                url = null,
-                key = null,
-            )
-            val mockFile = File.createTempFile("mockFile", "temp")
-            val mockAttachment = createMockSdkAttachment(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(userId = userId, cipherView = mockCipherView)
-            } returns mockCipher.asSuccess()
-            coEvery {
-                fileManager.writeUriToCache(fileUri = mockUri)
-            } returns mockFile.asSuccess()
-            coEvery {
-                vaultSdkSource.encryptAttachment(
-                    userId = userId,
-                    cipher = mockCipher,
-                    attachmentView = mockAttachmentView,
-                    decryptedFilePath = mockFile.absolutePath,
-                    encryptedFilePath = "${mockFile.absolutePath}.enc",
-                )
-            } returns mockAttachment.asSuccess()
-            coEvery {
-                ciphersService.createAttachment(
-                    cipherId = cipherId,
-                    body = AttachmentJsonRequest(
-                        fileName = mockFileName,
-                        key = "mockKey-1",
-                        fileSize = mockFileSize,
-                    ),
-                )
-            } returns Throwable("Fail").asFailure()
-
-            val result = vaultRepository.createAttachment(
-                cipherId = cipherId,
-                cipherView = mockCipherView,
-                fileSizeBytes = mockFileSize,
-                fileName = mockFileName,
-                fileUri = mockUri,
-            )
-
-            assertEquals(CreateAttachmentResult.Error, result)
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `createAttachment with uploadAttachment failure should return CreateAttachmentResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "cipherId-1"
-            val mockUri = setupMockUri(url = "www.test.com")
-            val mockCipherView = createMockCipherView(number = 1)
-            val mockCipher = createMockSdkCipher(number = 1, clock = clock)
-            val mockFileName = "mockFileName-1"
-            val mockFileSize = "1"
-            val mockAttachmentView = createMockAttachmentView(number = 1).copy(
-                sizeName = null,
-                id = null,
-                url = null,
-                key = null,
-            )
-            val mockFile = File.createTempFile("mockFile", "temp")
-            val mockAttachment = createMockSdkAttachment(number = 1)
-            val mockAttachmentJsonResponse = createMockAttachmentJsonResponse(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(userId = userId, cipherView = mockCipherView)
-            } returns mockCipher.asSuccess()
-            coEvery {
-                fileManager.writeUriToCache(fileUri = mockUri)
-            } returns mockFile.asSuccess()
-            coEvery {
-                vaultSdkSource.encryptAttachment(
-                    userId = userId,
-                    cipher = mockCipher,
-                    attachmentView = mockAttachmentView,
-                    decryptedFilePath = mockFile.absolutePath,
-                    encryptedFilePath = "${mockFile.absolutePath}.enc",
-                )
-            } returns mockAttachment.asSuccess()
-            coEvery {
-                ciphersService.createAttachment(
-                    cipherId = cipherId,
-                    body = AttachmentJsonRequest(
-                        fileName = mockFileName,
-                        key = "mockKey-1",
-                        fileSize = mockFileSize,
-                    ),
-                )
-            } returns mockAttachmentJsonResponse.asSuccess()
-            coEvery {
-                ciphersService.uploadAttachment(
-                    attachmentJsonResponse = mockAttachmentJsonResponse,
-                    encryptedFile = File("${mockFile.absoluteFile}.enc"),
-                )
-            } returns Throwable("Fail").asFailure()
-
-            val result = vaultRepository.createAttachment(
-                cipherId = cipherId,
-                cipherView = mockCipherView,
-                fileSizeBytes = mockFileSize,
-                fileName = mockFileName,
-                fileUri = mockUri,
-            )
-
-            assertEquals(CreateAttachmentResult.Error, result)
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `createAttachment with decryptCipher failure should return CreateAttachmentResult Error`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "cipherId-1"
-            val mockUri = setupMockUri(url = "www.test.com")
-            val mockCipherView = createMockCipherView(number = 1)
-            val mockCipher = createMockSdkCipher(number = 1, clock = clock)
-            val mockFileName = "mockFileName-1"
-            val mockFileSize = "1"
-            val mockAttachmentView = createMockAttachmentView(number = 1).copy(
-                sizeName = null,
-                id = null,
-                url = null,
-                key = null,
-            )
-            val mockFile = File.createTempFile("mockFile", "temp")
-            val mockAttachment = createMockSdkAttachment(number = 1)
-            val mockAttachmentJsonResponse = createMockAttachmentJsonResponse(number = 1)
-            val mockCipherResponse = createMockCipher(number = 1).copy(collectionIds = null)
-            val mockUpdatedCipherResponse = createMockCipher(number = 1).copy(
-                collectionIds = listOf("mockId-1"),
-            )
-            coEvery {
-                vaultSdkSource.encryptCipher(userId = userId, cipherView = mockCipherView)
-            } returns mockCipher.asSuccess()
-            coEvery {
-                fileManager.writeUriToCache(fileUri = mockUri)
-            } returns mockFile.asSuccess()
-            coEvery {
-                vaultSdkSource.encryptAttachment(
-                    userId = userId,
-                    cipher = mockCipher,
-                    attachmentView = mockAttachmentView,
-                    decryptedFilePath = mockFile.absolutePath,
-                    encryptedFilePath = "${mockFile.absolutePath}.enc",
-                )
-            } returns mockAttachment.asSuccess()
-            coEvery {
-                ciphersService.createAttachment(
-                    cipherId = cipherId,
-                    body = AttachmentJsonRequest(
-                        fileName = mockFileName,
-                        key = "mockKey-1",
-                        fileSize = mockFileSize,
-                    ),
-                )
-            } returns mockAttachmentJsonResponse.asSuccess()
-            coEvery {
-                ciphersService.uploadAttachment(
-                    attachmentJsonResponse = mockAttachmentJsonResponse,
-                    encryptedFile = File("${mockFile.absoluteFile}.enc"),
-                )
-            } returns mockCipherResponse.asSuccess()
-            coEvery {
-                vaultDiskSource.saveCipher(userId = userId, cipher = mockUpdatedCipherResponse)
-            } just runs
-            coEvery {
-                vaultSdkSource.decryptCipher(
-                    userId = userId,
-                    cipher = mockUpdatedCipherResponse.toEncryptedSdkCipher(),
-                )
-            } returns Throwable("Fail").asFailure()
-
-            val result = vaultRepository.createAttachment(
-                cipherId = cipherId,
-                cipherView = mockCipherView,
-                fileSizeBytes = mockFileSize,
-                fileName = mockFileName,
-                fileUri = mockUri,
-            )
-
-            assertEquals(CreateAttachmentResult.Error, result)
-        }
-
-    @Suppress("MaxLineLength")
-    @Test
-    fun `createAttachment with createAttachment success should return CreateAttachmentResult Success`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "cipherId-1"
-            val mockUri = setupMockUri(url = "www.test.com")
-            val mockCipherView = createMockCipherView(number = 1)
-            val mockCipher = createMockSdkCipher(number = 1, clock = clock)
-            val mockFileName = "mockFileName-1"
-            val mockFileSize = "1"
-            val mockAttachmentView = createMockAttachmentView(number = 1).copy(
-                sizeName = null,
-                id = null,
-                url = null,
-                key = null,
-            )
-            val mockFile = File.createTempFile("mockFile", "temp")
-            val mockAttachment = createMockSdkAttachment(number = 1)
-            val mockAttachmentJsonResponse = createMockAttachmentJsonResponse(number = 1)
-            val mockCipherResponse = createMockCipher(number = 1).copy(collectionIds = null)
-            val mockUpdatedCipherResponse = createMockCipher(number = 1).copy(
-                collectionIds = listOf("mockId-1"),
-            )
-            coEvery {
-                vaultSdkSource.encryptCipher(userId = userId, cipherView = mockCipherView)
-            } returns mockCipher.asSuccess()
-            coEvery {
-                fileManager.writeUriToCache(fileUri = mockUri)
-            } returns mockFile.asSuccess()
-            coEvery {
-                vaultSdkSource.encryptAttachment(
-                    userId = userId,
-                    cipher = mockCipher,
-                    attachmentView = mockAttachmentView,
-                    decryptedFilePath = mockFile.absolutePath,
-                    encryptedFilePath = "${mockFile.absolutePath}.enc",
-                )
-            } returns mockAttachment.asSuccess()
-            coEvery {
-                ciphersService.createAttachment(
-                    cipherId = cipherId,
-                    body = AttachmentJsonRequest(
-                        fileName = mockFileName,
-                        key = "mockKey-1",
-                        fileSize = mockFileSize,
-                    ),
-                )
-            } returns mockAttachmentJsonResponse.asSuccess()
-            coEvery {
-                ciphersService.uploadAttachment(
-                    attachmentJsonResponse = mockAttachmentJsonResponse,
-                    encryptedFile = File("${mockFile.absolutePath}.enc"),
-                )
-            } returns mockCipherResponse.asSuccess()
-            coEvery {
-                vaultDiskSource.saveCipher(userId = userId, cipher = mockUpdatedCipherResponse)
-            } just runs
-            coEvery {
-                vaultSdkSource.decryptCipher(
-                    userId = userId,
-                    cipher = mockUpdatedCipherResponse.toEncryptedSdkCipher(),
-                )
-            } returns mockCipherView.asSuccess()
-
-            val result = vaultRepository.createAttachment(
-                cipherId = cipherId,
-                cipherView = mockCipherView,
-                fileSizeBytes = mockFileSize,
-                fileName = mockFileName,
-                fileUri = mockUri,
-            )
-
-            assertEquals(CreateAttachmentResult.Success(mockCipherView), result)
-        }
-
-    @Test
-    fun `createAttachment should delete temp files after upload success`() {
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "cipherId-1"
-            val mockUri = setupMockUri(url = "www.test.com")
-            val mockCipherView = createMockCipherView(number = 1)
-            val mockCipher = createMockSdkCipher(number = 1, clock = clock)
-            val mockFileName = "mockFileName-1"
-            val mockFileSize = "1"
-            val mockAttachmentView = createMockAttachmentView(number = 1).copy(
-                sizeName = null,
-                id = null,
-                url = null,
-                key = null,
-            )
-            val mockFile = File.createTempFile("mockFile", "temp")
-            val mockAttachment = createMockSdkAttachment(number = 1)
-            val mockAttachmentJsonResponse = createMockAttachmentJsonResponse(number = 1)
-            val mockCipherResponse = createMockCipher(number = 1).copy(collectionIds = null)
-            val mockUpdatedCipherResponse = createMockCipher(number = 1).copy(
-                collectionIds = listOf("mockId-1"),
-            )
-            coEvery {
-                vaultSdkSource.encryptCipher(userId = userId, cipherView = mockCipherView)
-            } returns mockCipher.asSuccess()
-            coEvery {
-                fileManager.writeUriToCache(fileUri = mockUri)
-            } returns mockFile.asSuccess()
-            coEvery {
-                vaultSdkSource.encryptAttachment(
-                    userId = userId,
-                    cipher = mockCipher,
-                    attachmentView = mockAttachmentView,
-                    decryptedFilePath = mockFile.absolutePath,
-                    encryptedFilePath = "${mockFile.absolutePath}.enc",
-                )
-            } returns mockAttachment.asSuccess()
-            coEvery {
-                ciphersService.createAttachment(
-                    cipherId = cipherId,
-                    body = AttachmentJsonRequest(
-                        fileName = mockFileName,
-                        key = "mockKey-1",
-                        fileSize = mockFileSize,
-                    ),
-                )
-            } returns mockAttachmentJsonResponse.asSuccess()
-            coEvery {
-                ciphersService.uploadAttachment(
-                    attachmentJsonResponse = mockAttachmentJsonResponse,
-                    encryptedFile = File("${mockFile.absolutePath}.enc"),
-                )
-            } returns mockCipherResponse.asSuccess()
-            coEvery {
-                vaultDiskSource.saveCipher(userId = userId, cipher = mockUpdatedCipherResponse)
-            } just runs
-            coEvery {
-                vaultSdkSource.decryptCipher(
-                    userId = userId,
-                    cipher = mockUpdatedCipherResponse.toEncryptedSdkCipher(),
-                )
-            } returns mockCipherView.asSuccess()
-
-            vaultRepository.createAttachment(
-                cipherId = cipherId,
-                cipherView = mockCipherView,
-                fileSizeBytes = mockFileSize,
-                fileName = mockFileName,
-                fileUri = mockUri,
-            )
-
-            coVerify {
-                fileManager.delete(*anyVararg())
-            }
-        }
-    }
-
-    @Test
-    fun `createAttachment should delete temp files after upload failure`() {
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-            val userId = "mockId-1"
-            val cipherId = "cipherId-1"
-            val mockUri = setupMockUri(url = "www.test.com")
-            val mockCipherView = createMockCipherView(number = 1)
-            val mockCipher = createMockSdkCipher(number = 1, clock = clock)
-            val mockFileName = "mockFileName-1"
-            val mockFileSize = "1"
-            val mockAttachmentView = createMockAttachmentView(number = 1).copy(
-                sizeName = null,
-                id = null,
-                url = null,
-                key = null,
-            )
-            val mockFile = File.createTempFile("mockFile", "temp")
-            val mockAttachment = createMockSdkAttachment(number = 1)
-            val mockAttachmentJsonResponse = createMockAttachmentJsonResponse(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(userId = userId, cipherView = mockCipherView)
-            } returns mockCipher.asSuccess()
-            coEvery {
-                fileManager.writeUriToCache(fileUri = mockUri)
-            } returns mockFile.asSuccess()
-            coEvery {
-                vaultSdkSource.encryptAttachment(
-                    userId = userId,
-                    cipher = mockCipher,
-                    attachmentView = mockAttachmentView,
-                    decryptedFilePath = mockFile.absolutePath,
-                    encryptedFilePath = "${mockFile.absolutePath}.enc",
-                )
-            } returns mockAttachment.asSuccess()
-            coEvery {
-                ciphersService.createAttachment(
-                    cipherId = cipherId,
-                    body = AttachmentJsonRequest(
-                        fileName = mockFileName,
-                        key = "mockKey-1",
-                        fileSize = mockFileSize,
-                    ),
-                )
-            } returns mockAttachmentJsonResponse.asSuccess()
-            coEvery {
-                ciphersService.uploadAttachment(
-                    attachmentJsonResponse = mockAttachmentJsonResponse,
-                    encryptedFile = File("${mockFile.absolutePath}.enc"),
-                )
-            } returns Throwable("Fail").asFailure()
-
-            vaultRepository.createAttachment(
-                cipherId = cipherId,
-                cipherView = mockCipherView,
-                fileSizeBytes = mockFileSize,
-                fileName = mockFileName,
-                fileUri = mockUri,
-            )
-
-            coVerify {
-                fileManager.delete(*anyVararg())
-            }
-        }
-    }
-
-    @Test
-    fun `downloadAttachment with missing attachment should return Failure`() = runTest {
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-
-        val attachmentId = "mockId-1"
-        val cipher = mockk<Cipher> {
-            every { attachments } returns emptyList()
-            every { id } returns "mockId-1"
-        }
-        val cipherView = createMockCipherView(number = 1)
-        coEvery {
-            vaultSdkSource.encryptCipher(MOCK_USER_STATE.activeUserId, cipherView)
-        } returns cipher.asSuccess()
-
-        assertEquals(
-            DownloadAttachmentResult.Failure,
-            vaultRepository.downloadAttachment(
-                cipherView = cipherView,
-                attachmentId = attachmentId,
-            ),
-        )
-
-        coVerify(exactly = 1) {
-            vaultSdkSource.encryptCipher(MOCK_USER_STATE.activeUserId, cipherView)
-        }
-        coVerify(exactly = 0) {
-            ciphersService.getCipherAttachment(any(), any())
-        }
-    }
-
-    @Test
-    fun `downloadAttachment with failed attachment details request should return Failure`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-
-            val attachmentId = "mockId-1"
-            val attachment = mockk<Attachment> {
-                every { id } returns attachmentId
-            }
-            val cipher = mockk<Cipher> {
-                every { attachments } returns listOf(attachment)
-                every { id } returns "mockId-1"
-            }
-            val cipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(MOCK_USER_STATE.activeUserId, cipherView)
-            } returns cipher.asSuccess()
-
-            coEvery {
-                ciphersService.getCipherAttachment(any(), any())
-            } returns Throwable().asFailure()
-
-            assertEquals(
-                DownloadAttachmentResult.Failure,
-                vaultRepository.downloadAttachment(
-                    cipherView = cipherView,
-                    attachmentId = attachmentId,
-                ),
-            )
-
-            coVerify(exactly = 1) {
-                vaultSdkSource.encryptCipher(MOCK_USER_STATE.activeUserId, cipherView)
-                ciphersService.getCipherAttachment(
-                    cipherId = requireNotNull(cipherView.id),
-                    attachmentId = attachmentId,
-                )
-            }
-        }
-
-    @Test
-    fun `downloadAttachment with attachment details missing url should return Failure`() = runTest {
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-
-        val attachmentId = "mockId-1"
-        val attachment = mockk<Attachment> {
-            every { id } returns attachmentId
-        }
-        val cipher = mockk<Cipher> {
-            every { attachments } returns listOf(attachment)
-            every { id } returns "mockId-1"
-        }
-        val cipherView = createMockCipherView(number = 1)
-        coEvery {
-            vaultSdkSource.encryptCipher(MOCK_USER_STATE.activeUserId, cipherView)
-        } returns cipher.asSuccess()
-
-        val response = mockk<SyncResponseJson.Cipher.Attachment> {
-            every { url } returns null
-        }
-        coEvery {
-            ciphersService.getCipherAttachment(any(), any())
-        } returns response.asSuccess()
-
-        assertEquals(
-            DownloadAttachmentResult.Failure,
-            vaultRepository.downloadAttachment(
-                cipherView = cipherView,
-                attachmentId = attachmentId,
-            ),
-        )
-
-        coVerify(exactly = 1) {
-            vaultSdkSource.encryptCipher(MOCK_USER_STATE.activeUserId, cipherView)
-            ciphersService.getCipherAttachment(
-                cipherId = requireNotNull(cipherView.id),
-                attachmentId = attachmentId,
-            )
-        }
-    }
-
-    @Test
-    fun `downloadAttachment with failed download should return Failure`() = runTest {
-        fakeAuthDiskSource.userState = MOCK_USER_STATE
-
-        val attachmentId = "mockId-1"
-        val attachment = mockk<Attachment> {
-            every { id } returns attachmentId
-        }
-        val cipher = mockk<Cipher> {
-            every { attachments } returns listOf(attachment)
-            every { id } returns "mockId-1"
-        }
-
-        val cipherView = createMockCipherView(number = 1)
-        coEvery {
-            vaultSdkSource.encryptCipher(MOCK_USER_STATE.activeUserId, cipherView)
-        } returns cipher.asSuccess()
-
-        val response = mockk<SyncResponseJson.Cipher.Attachment> {
-            every { url } returns "https://bitwarden.com"
-        }
-        coEvery {
-            ciphersService.getCipherAttachment(any(), any())
-        } returns response.asSuccess()
-
-        coEvery {
-            fileManager.downloadFileToCache(any())
-        } returns DownloadResult.Failure
-
-        assertEquals(
-            DownloadAttachmentResult.Failure,
-            vaultRepository.downloadAttachment(
-                cipherView = cipherView,
-                attachmentId = attachmentId,
-            ),
-        )
-
-        coVerify(exactly = 1) {
-            vaultSdkSource.encryptCipher(MOCK_USER_STATE.activeUserId, cipherView)
-            ciphersService.getCipherAttachment(
-                cipherId = requireNotNull(cipherView.id),
-                attachmentId = attachmentId,
-            )
-            fileManager.downloadFileToCache("https://bitwarden.com")
-        }
-    }
-
-    @Test
-    fun `downloadAttachment with failed decryption should delete file and return Failure`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-
-            val attachmentId = "mockId-1"
-            val attachment = mockk<Attachment> {
-                every { id } returns attachmentId
-            }
-            val cipher = mockk<Cipher> {
-                every { attachments } returns listOf(attachment)
-                every { id } returns "mockId-1"
-            }
-            val cipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(MOCK_USER_STATE.activeUserId, cipherView)
-            } returns cipher.asSuccess()
-
-            val response = mockk<SyncResponseJson.Cipher.Attachment> {
-                every { url } returns "https://bitwarden.com"
-            }
-            coEvery {
-                ciphersService.getCipherAttachment(any(), any())
-            } returns response.asSuccess()
-
-            val file = mockk<File> {
-                every { path } returns "path/to/encrypted/file"
-                every { delete() } returns true
-            }
-            coEvery {
-                fileManager.downloadFileToCache(any())
-            } returns DownloadResult.Success(file)
-
-            coEvery {
-                vaultSdkSource.decryptFile(
-                    userId = MOCK_USER_STATE.activeUserId,
-                    cipher = cipher,
-                    attachment = attachment,
-                    encryptedFilePath = "path/to/encrypted/file",
-                    decryptedFilePath = "path/to/encrypted/file_decrypted",
-                )
-            } returns Throwable().asFailure()
-
-            assertEquals(
-                DownloadAttachmentResult.Failure,
-                vaultRepository.downloadAttachment(
-                    cipherView = cipherView,
-                    attachmentId = attachmentId,
-                ),
-            )
-
-            coVerify(exactly = 1) {
-                vaultSdkSource.encryptCipher(MOCK_USER_STATE.activeUserId, cipherView)
-                ciphersService.getCipherAttachment(
-                    cipherId = requireNotNull(cipherView.id),
-                    attachmentId = attachmentId,
-                )
-                fileManager.downloadFileToCache("https://bitwarden.com")
-                vaultSdkSource.decryptFile(
-                    userId = MOCK_USER_STATE.activeUserId,
-                    cipher = cipher,
-                    attachment = attachment,
-                    encryptedFilePath = "path/to/encrypted/file",
-                    decryptedFilePath = "path/to/encrypted/file_decrypted",
-                )
-            }
-            verify(exactly = 1) {
-                file.delete()
-            }
-        }
-
-    @Test
-    fun `downloadAttachment with successful decryption should delete file and return Success`() =
-        runTest {
-            fakeAuthDiskSource.userState = MOCK_USER_STATE
-
-            val attachmentId = "mockId-1"
-            val attachment = mockk<Attachment> {
-                every { id } returns attachmentId
-            }
-            val cipher = mockk<Cipher> {
-                every { attachments } returns listOf(attachment)
-                every { id } returns "mockId-1"
-            }
-            val cipherView = createMockCipherView(number = 1)
-            coEvery {
-                vaultSdkSource.encryptCipher(MOCK_USER_STATE.activeUserId, cipherView)
-            } returns cipher.asSuccess()
-
-            val response = mockk<SyncResponseJson.Cipher.Attachment> {
-                every { url } returns "https://bitwarden.com"
-            }
-            coEvery {
-                ciphersService.getCipherAttachment(any(), any())
-            } returns response.asSuccess()
-
-            val file = mockk<File> {
-                every { path } returns "path/to/encrypted/file"
-                every { delete() } returns true
-            }
-            coEvery {
-                fileManager.downloadFileToCache(any())
-            } returns DownloadResult.Success(file)
-
-            coEvery {
-                vaultSdkSource.decryptFile(
-                    userId = MOCK_USER_STATE.activeUserId,
-                    cipher = cipher,
-                    attachment = attachment,
-                    encryptedFilePath = "path/to/encrypted/file",
-                    decryptedFilePath = "path/to/encrypted/file_decrypted",
-                )
-            } returns Unit.asSuccess()
-
-            assertEquals(
-                DownloadAttachmentResult.Success(
-                    file = File("path/to/encrypted/file_decrypted"),
-                ),
-                vaultRepository.downloadAttachment(
-                    cipherView = cipherView,
-                    attachmentId = attachmentId,
-                ),
-            )
-
-            coVerify(exactly = 1) {
-                vaultSdkSource.encryptCipher(MOCK_USER_STATE.activeUserId, cipherView)
-                ciphersService.getCipherAttachment(
-                    cipherId = requireNotNull(cipherView.id),
-                    attachmentId = attachmentId,
-                )
-                fileManager.downloadFileToCache("https://bitwarden.com")
-                vaultSdkSource.decryptFile(
-                    userId = MOCK_USER_STATE.activeUserId,
-                    cipher = cipher,
-                    attachment = attachment,
-                    encryptedFilePath = "path/to/encrypted/file",
-                    decryptedFilePath = "path/to/encrypted/file_decrypted",
-                )
-            }
-            verify(exactly = 1) {
-                file.delete()
-            }
         }
 
     @Test
